Bugfixes, new ripple event search, and encode parallelization breakout.
New Features:
* Event data containers events can be search for by timestamp.

Notebooks:
* Updated encoding algorithm notebook to decode more tetrodes of bond.
* Encoding notebook breaks out an exmaple of encoding parallelization
strategy.

Bugfixes:
* Linear position data containers create_default did not enforce
sampling_rate be set as parameter.
* In data generator example notebook, when switching to the encoding
class, the encoding spikes was being coverted to floats and passed to
normal_pdf_int_lookup because in setting the velocity threshold,
requiring a resampling of linear position, the sampling_rate was not
set.  This caused the time index to be set to inf, which converted
spk_amp data to floats.