Fixed missing_bins bug in decoder.
Decoder using the 'missing_bins' field of the observation to determine
when 0 spikes have fired in a bin.  'missing_bins' is only assigned to
the last electrode in the current bin, all other observations on other
electrodes must have 0 for 'missing_bins'.  But sometimes the order of
the spikes can change if the spikes occurred at the same time (of course
on different electrodes).  Then means the last spike's 'missing_bin'
does not have to have the correct number, because order when parsing the
data for parallel computations is not guarenteed.

There is probably a better way to track bins with 0 spikes.