Standardized visualization functions for posteriors.
Fixed inconsistent parameters for each plot function.
For plot functions default x-axis is the entire posterior.
When time and plt_range are specified, they are used.
When x_range and y_range are specified, then they are used
and override time and plt_range.

Added parameter for max value of heatmap.